move DEFAULT_STATE to types where it can be appropriately shared between extension and webview
