+++ editors/rift-vscode-svelte/webviews/components/stores.ts
@@ -1,21 +1,8 @@
 import { readable, writable } from 'svelte/store'
-// import { DEFAULT_STATE } from '../../src/client'
+import { DEFAULT_STATE } from '../../src/types'
 import type { WebviewState } from '../../src/types'
 
 
-const DEFAULT_STATE: WebviewState = {
-  selectedAgentId: '',
-  agents: {
-  },
-  availableAgents: [{
-    agent_type: "rift_chat",
-    agent_description: '',
-    agent_icon: '',
-    display_name: 'Rift Chat'
-  }],
-  isStreaming: false,
-  streamingText: ''
-}
 
 export const state = readable<WebviewState>(DEFAULT_STATE, (set) => {
   window.addEventListener('message', (event:any) => {
